+++ src/cppwasm/MathUtil.hpp
@@ -13,9 +13,9 @@ namespace constantq
     public:
         static unsigned int leadingZeros(unsigned int x);
         static unsigned int reverse(unsigned int num);
-        static void fft(std::vector<std::complex<double>> &x, int n);
+        static void fft(std::vector<std::complex<double>> &x, size_t n);
         static int nextPow2(double num);
-        static std::vector<std::complex<double>> hamming(int len);
+        static std::vector<std::complex<double>> hamming(size_t len);
         static std::complex<double> eulers(double num);
     };
 }
